Partial refactor of db model structure, for #419
