+++ app/db_models/user_data.py
@@ -2,11 +2,9 @@ from google.appengine.ext import db
 import pprint
 import sys
 import os
-sys.path.append(os.path.join(os.path.dirname(__file__), 'backend' ))
-#import backendservice
 from backend.common.config import AgentTypes, JobConfig, JobDatabaseConfig
 import logging
-from vm_state_model import VMStateModel
+from db_models.vm_state_model import VMStateModel
 
 try:
     import json
