+++ app/handlers/stochoptim.py
@@ -26,6 +26,10 @@ import traceback
 from backend.backendservice import backendservices
 from backend.common.config import AgentTypes, JobConfig
 
+from db_models.stochoptim_job import StochOptimJobWrapper
+
+
+
 def int_or_float(s):
     try:
         return int(s)
@@ -149,72 +153,6 @@ class StochOptimModel(stochss.model.Model):
         else:
             return os.linesep.join([rnu, pnu, snames, rnames, rparms, rknames, rconstant, rkind])
 
-class StochOptimJobWrapper(db.Model):
-
-    userId = db.StringProperty()
-    pid = db.IntegerProperty()
-    startTime = db.StringProperty()
-    jobName = db.StringProperty()
-    modelName = db.StringProperty()
-    indata = db.TextProperty()
-    nameToIndex = db.TextProperty()
-    outData = db.StringProperty()
-    status = db.StringProperty()
-    zipFileName = db.StringProperty()
-    
-    resource = db.StringProperty()
-    outputURL = db.StringProperty()
-    cloudDatabaseID = db.StringProperty()
-    celeryPID = db.StringProperty()
-    pollProcessPID = db.IntegerProperty()
-
-    def delete(self, handler):
-        logging.debug("StochOptimJobWrapper(cloudDatabaseID={0})".format(self.cloudDatabaseID))
-        if self.outData is not None and os.path.exists(self.outData):
-            shutil.rmtree(self.outData)
-        
-        if self.zipFileName is not None and os.path.exists(self.zipFileName):
-                os.remove(self.zipFileName)
-
-        self.stop(handler)
-        
-        # delete on cloud
-        if self.resource in backendservices.SUPPORTED_CLOUD_RESOURCES:
-            service = backend.backendservice.backendservices(handler.user_data)
-            service.deleteTasks(self)
-
-        super(StochOptimJobWrapper, self).delete()
-
-    def stop(self, handler):
-        if self.status == "Running" or self.status == "Pending":
-            service = backend.backendservice.backendservices(handler.user_data)
-            if self.resource.lower() == "local":
-                service.deleteTaskLocal([int(self.pid)])
-            elif self.resource in backendservices.SUPPORTED_CLOUD_RESOURCES:
-                result = service.stopTasks(self)
-                if result and result[self.cloudDatabaseID]:
-                    final_cloud_result = result[self.cloudDatabaseID]
-                    try:
-                        self.outputURL = final_cloud_result['output']
-                    except KeyError:
-                        pass
-                    self.status = "Finished"
-                    self.put()
-                    return True
-                else:
-                    # Something went wrong
-                    logging.error(result)
-                    return False
-            else:
-                raise Exception("Unknown job resource '{0}'".format(sef.resource))
-    
-    def mark_final_cloud_data(self):
-        flag_file = os.path.join(self.outData, ".final-cloud")
-        os.system("touch {0}".format(flag_file))
-    
-    def has_final_cloud_data(self):
-        flag_file = os.path.join(self.outData, ".final-cloud")
-        return os.path.exists(flag_file)
 
 class StochOptimPage(BaseHandler):
     def authentication_required(self):
@@ -464,7 +402,10 @@ class StochOptimPage(BaseHandler):
         jFileData = fileserver.FileManager.getFile(self, data["trajectoriesID"], noFile = False)
         iFileData = fileserver.FileManager.getFile(self, data["initialDataID"], noFile = False)
 
+        job.put()
+
         cloud_params = {
+            "job_id" : job.key().id(),
             "job_type": "mcem2",
             # "cores": data["cores"],
             "paramstring": cmd,
@@ -717,8 +658,8 @@ class StochOptimVisualization(BaseHandler):
             result['status']=True
             result['msg'] = "Successfully fetched the remote output files."
 
-        except Exception,e:
-            logging.info('************************************* {0}'.format(e))
+        except Exception as e:
+            logging.info('StochOptim: Failed to fetch the remote files. {0}'.format(e))
             result['status']=False
             result['msg'] = "Failed to fetch the remote files."
         return result
